refactor(decoration): encapsulate module-level state into DecorationService class
Replace global state (output, decorationTypes, documentCaches) with a
class-based approach implementing vscode.Disposable for proper lifecycle
management.
