--- src/decoration/clearDocumentCache.ts
@@ -1,5 +1,0 @@
-import { documentCaches } from './state'
-
-export function clearDocumentCache(uri: string): void {
-  documentCaches.delete(uri)
-}
