--- src/decoration/disposeDecorations.ts
@@ -1,9 +1,0 @@
-import { decorationTypes, documentCaches } from './state'
-
-export function disposeDecorations(): void {
-  for (const type of decorationTypes.values()) {
-    type.dispose()
-  }
-  decorationTypes.clear()
-  documentCaches.clear()
-}
