+++ src/extension.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode'
-import { applyImportDecorations, clearDocumentCache, disposeDecorations } from './decoration'
+import { DecorationService } from './decoration'
 
 const SUPPORTED_LANGUAGES = new Set(['typescript', 'typescriptreact'])
 
@@ -7,13 +7,15 @@ function isSupported(document: vscode.TextDocument): boolean {
   return document.uri.scheme === 'file' && SUPPORTED_LANGUAGES.has(document.languageId)
 }
 
-function triggerDecoration(editor: vscode.TextEditor): void {
-  applyImportDecorations(editor).catch(() => {
-    // TS language service may not be ready yet; silently ignore
-  })
-}
-
 export function activate(context: vscode.ExtensionContext): void {
+  const service = new DecorationService()
+
+  function triggerDecoration(editor: vscode.TextEditor): void {
+    service.applyImportDecorations(editor).catch(() => {
+      // TS language service may not be ready yet; silently ignore
+    })
+  }
+
   for (const editor of vscode.window.visibleTextEditors) {
     if (isSupported(editor.document)) {
       triggerDecoration(editor)
@@ -21,6 +23,7 @@ export function activate(context: vscode.ExtensionContext): void {
   }
 
   context.subscriptions.push(
+    service,
     vscode.window.onDidChangeActiveTextEditor((editor) => {
       if (editor && isSupported(editor.document)) {
         triggerDecoration(editor)
@@ -34,12 +37,12 @@ export function activate(context: vscode.ExtensionContext): void {
     }),
     vscode.workspace.onDidCloseTextDocument((document) => {
       if (isSupported(document)) {
-        clearDocumentCache(document.uri.toString())
+        service.clearDocumentCache(document.uri.toString())
       }
     }),
   )
 }
 
 export function deactivate(): void {
-  disposeDecorations()
+  // Service is disposed automatically via context.subscriptions
 }
